Refactor Folia support into Spigot implementation
