--- src/main/resources/paper-plugin.yml
@@ -1,6 +1,0 @@
-name: Pano
-main: com.panomc.plugins.pano.spigot.SpigotMain
-version: ${version}
-api-version: '1.20'
-folia-supported: true
-folia-main: com.panomc.plugins.pano.folia.FoliaMain
